refactor: move button-bar, button-group, and toolbar to actions
