+++ src/components/actions/button-group/rr-button-group.stories.js
@@ -1,9 +1,9 @@
 import { html } from 'lit';
 import './rr-button-group.ts';
-import '../../actions/button/rr-button.ts';
+import '../button/rr-button.ts';
 
 export default {
-  title: 'Components/Control Groups/Button Group',
+  title: 'Components/Actions/Button Group',
   component: 'rr-button-group',
   tags: ['autodocs'],
   parameters: {
